refactor cost param estimation
